refactor(static-website): move core logic to abstract class
